Safari-611.1.10/LayoutTests/accessibility/ios-simulator/definition-list-expected.txt

term 1
def 1
This tests that the definition lists are exposed on ios.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS term.isInDefinitionListTerm is true
PASS def.isInDefinitionListDefinition is true
PASS successfullyParsed is true

TEST COMPLETE
